requirements.txt
absl-py==1.4.0
accelerate==0.17.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
bleach==6.0.0
blinker==1.4
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.0
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
datasets==2.10.1
dbus-python==1.2.18
dill==0.3.6
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
docutils==0.20.1
filelock==3.12.2
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
gyp==0.1
httplib2==0.20.2
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.0
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
monai==1.1.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
netifaces==0.11.0
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==3.8.0
protobuf==4.22.1
psutil==5.9.4
py-brainage==0.0.7
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
readme-renderer==40.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
rfc3986==2.0.0
rich==13.4.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.1
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
sympy==1.11.1
systemd-python==234
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.2
triton==2.0.0
twine==4.0.2
typing_extensions==4.5.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.15
virtualenv==20.23.1
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==1.0.0
